LayoutTests/fast/css/pseudo-class-internal-expected.txt

This tests that -internal- pseudo classes are not exposed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS target.matches(":-internal-direct-focus") threw exception SyntaxError: ':-internal-direct-focus' is not a valid selector..
PASS target.matches(":-internal-html-document") threw exception SyntaxError: ':-internal-html-document' is not a valid selector..
PASS successfullyParsed is true

TEST COMPLETE
